Moved measure to higher level and renamed measure file to solvers
