+++ mat_dp_core/__init__.py
@@ -7,9 +7,9 @@ from mat_dp_core.maths_core import (
     EqConstraint,
     GeConstraint,
     LeConstraint,
-    Measure,
     Process,
     Processes,
     Resource,
     Resources,
 )
+from mat_dp_core.measure import Measure
