+++ mat_dp_core/maths_core/__init__.py
@@ -1,4 +1,4 @@
 from .constraints import EqConstraint, GeConstraint, LeConstraint
-from .measure import Measure
 from .processes import Process, Processes
 from .resources import Resource, Resources
+from .solvers import BoundedSolver
